(#8758) Refactor 'yumrepo' to use a provider
The current version of 'yumrepo' does not use a
provider, and it's one of only a couple of types
like that.  This patch refactors the type to use
a provider, and it does so in a relatively simple
way: It essentially moves a bunch of code to
a provider class, and then s/resource/provider/g.

Unfortunately there are no tests at all for yumrepo,
and this patch does not create any.

Signed-off-by: Luke Kanies <luke@puppetlabs.com>