Adding support for local resource auditing
This could be useful when a user is making changes
to a resource and wants to tell the sysadmin about
those changes.

Again, a rough prototype, but at least it shows what
the work would look like.

There's a lot of UX left to do, especially around tracking
resources that are being audited because of local
configurations rather than because they're in
the central catalog.

Signed-off-by: Luke Kanies <luke@puppetlabs.com>